fix: decouple format sentences from google cloud fns
